requirements.txt
Django==1.8.2
Markdown==2.6.2
Pillow==2.9.0
django-filer==0.9.11
django-pipeline==1.5.2
djangocms-admin-style==0.2.7
psycopg2==2.6.1
wsgiref==0.1.2
